Backend API Working,  Login Issue - Working, not clear.

+++ backend/controllers/foodController.js
@@ -1,8 +1,13 @@
-import foodModel from "../models/foodModels.js";
+import foodModel from "../models/foodModel.js";
 import fs from 'fs';
+
 
 // Add food item
 const addFood = async (req,res) => {
+
+    if (!req.file) {
+        return res.status(400).json({ success: false, message: "Nenhuma imagem enviada." });
+    }
 
     let image_filename = `${req.file.filename}`;
 
@@ -16,10 +21,10 @@ const addFood = async (req,res) => {
 
     try {
         await food.save();
-        res.json({sucess:true,message:"- #LOG: Food Added"})
+        res.json({success:true,message:"- #LOG: Food Added"})
     } catch (error){
         console.log(error)
-        res.json({sucess:false,message:"- #LOG-ERROR: Error in the const 'addFood', try-catch method. (File: '../backend/controllers/foodController.js'."})
+        res.json({success:false,message:"- #LOG-ERROR: Error in the const 'addFood', try-catch method. (File: '../backend/controllers/foodController.js'."})
     }
 
 }
@@ -28,10 +33,10 @@ const addFood = async (req,res) => {
 const listFood = async(req,res) => {
     try {
         const foods = await foodModel.find({});
-        res.json({sucess:true,data:foods})
+        res.json({success:true,data:foods})
     } catch (error){
         console.log(error);
-        res.json({sucess:false,message:"- #LOG-ERROR: Error in the const 'listFood', try-catch method. (File: '../backend/controllers/foodController.js'."})
+        res.json({success:false,message:"- #LOG-ERROR: Error in the const 'listFood', try-catch method. (File: '../backend/controllers/foodController.js'."})
     }
 }
 
@@ -39,13 +44,16 @@ const listFood = async(req,res) => {
 const removeFood = async(req,res) => {
     try{
         const food = await foodModel.findById(req.body.id);
-        fs.unlink(`uploads/${food.image}`,()=>{})
+        if (!food) {
+            return res.status(404).json({ success: false, message: "Comida não encontrada." });
+        }
+        fs.unlink(`uploads/${food.image}`,()=>{});
 
         await foodModel.findByIdAndDelete(req.body.id);
-        res.json({sucess:true,message:"- #LOG: Food removed!"})
+        res.json({success:true,message:"- #LOG: Food removed!"})
     } catch (error){
         console.log(error);
-        res.json({sucess:false,message:"- #LOG-ERROR: Error in the const 'removeFood' in the file '../backend/controllers/foodController.js'."})
+        res.json({success:false,message:"- #LOG-ERROR: Error in the const 'removeFood' in the file '../backend/controllers/foodController.js'."})
     }
 }
